+++ R/utilities.R
@@ -35,9 +35,9 @@ download = function(url, destfile, n = 3L) {
 
 
 connect = function(dbtype, dbname, ...) {
-  dbtype = match.arg(dbtype, c('postgres', 'mariadb', 'mysql', 'sqlite',
-                               'bigquery'))
-                               # 'clickhouse', 'bigquery'))
+  dbtype = match.arg(
+    dbtype, c('postgres', 'mariadb', 'mysql', 'sqlite', 'bigquery'))
+
   pkgName = switch(dbtype,
                    postgres = 'RPostgres',
                    mariadb = 'RMariaDB',
@@ -111,7 +111,8 @@ getParseFuncs = function(steps = 'all') {
     comment = parseComment,
     abstract = parseAbstract,
     author = parseAuthor,
-    investigator = parseInvestigator)
+    investigator = parseInvestigator,
+    other = parseOther)
 
   x = if ('all' %in% steps) parseFuncs else
     parseFuncs[names(parseFuncs) %in% steps]