+++ NEWS.md
@@ -1,3 +1,9 @@
+# pmparser 1.0.13
+* Added tables from the OtherAbstract and OtherID sections.
+
+# pmparser 1.0.12
+* Fixed issues found by lintr.
+
 # pmparser 1.0.11
 * Updated test standards based on PubMed's updated README.txt.